chore: move delete buttons back to sidebar
