+++ src/components/BottomBar.tsx
@@ -1,13 +1,9 @@
-import { useState, useCallback, useEffect, SetStateAction } from 'react'
+import { useState, useCallback, useEffect } from 'react'
 import { marked } from 'marked'
-import {
-  TrashIcon,
-  PencilSquareIcon,
-  RssIcon,
-} from '@heroicons/react/24/outline'
+import { RssIcon } from '@heroicons/react/24/outline'
 import { api } from '../state/api'
 import { useStore } from '../state/base'
-import { Publish, ConfirmDeleteDraft, ConfirmUnpublish } from './Modal'
+import { Publish } from './Modal'
 
 type BottomBarProps = {
   disabled: boolean
@@ -24,8 +20,6 @@ export default function BottomBar({ fileName, disabled }: BottomBarProps) {
     setActiveTheme,
   } = useStore()
 
-  const [showDeleteDraftModal, setShowDeleteDraftModal] = useState(false)
-  const [showUnpublishModal, setShowUnpublishModal] = useState(false)
   const [showPublishModal, setShowPublishModal] = useState(false)
 
   useEffect(() => {
@@ -58,33 +52,9 @@ export default function BottomBar({ fileName, disabled }: BottomBarProps) {
     [fileName, markdown, activeTheme]
   )
 
-  const handleDeleteDraft = useCallback(async (toDelete: string) => {
-    await api.poke({
-      app: 'blog',
-      mark: 'blog-action',
-      json: {
-        'delete-draft': {
-          path: toDelete,
-        },
-      },
-    })
-    setShowDeleteDraftModal(false)
-    getAll()
-  }, [])
-
-  const handleUnpublish = useCallback(async (toUnpublish: string) => {
-    await api.poke({
-      app: 'blog',
-      mark: 'blog-action',
-      json: { unpublish: { path: toUnpublish } },
-    })
-    setShowUnpublishModal(false)
-    getAll()
-  }, [])
-
   return (
     <>
-      <div className='w-1/4'>
+      <div className='w-1/2'>
         %theme:
         <select
           className='rounded border-none focus:outline-none'
@@ -99,30 +69,6 @@ export default function BottomBar({ fileName, disabled }: BottomBarProps) {
         </select>
       </div>
       <button
-        className='flex-1 flex items-center justify-center text-red-500 hover:text-white hover:bg-red-500 border rounded disabled:opacity-50'
-        disabled={!fileName}
-        onClick={() => {
-          setShowDeleteDraftModal(true)
-        }}
-      >
-        <div className='w-5 mr-2'>
-          <TrashIcon />
-        </div>
-        <code>delete %blog</code>
-      </button>
-      <button
-        className='flex-1 flex items-center justify-center text-white bg-orange-500 hover:bg-orange-700 rounded disabled:opacity-50'
-        disabled={!fileName || disabled}
-        onClick={() => {
-          setShowUnpublishModal(true)
-        }}
-      >
-        <div className='w-5 mr-2'>
-          <PencilSquareIcon />
-        </div>
-        <code>unpublish %blog</code>
-      </button>
-      <button
         className='flex-1 flex items-center justify-center text-white bg-green-500 hover:bg-green-700 rounded disabled:opacity-50'
         disabled={!fileName || disabled}
         onClick={handlePublish}
@@ -132,20 +78,6 @@ export default function BottomBar({ fileName, disabled }: BottomBarProps) {
         </div>
         <code>publish %blog</code>
       </button>
-      {showDeleteDraftModal && (
-        <ConfirmDeleteDraft
-          fileName={fileName}
-          setShowModal={setShowDeleteDraftModal}
-          onConfirm={() => handleDeleteDraft(fileName)}
-        />
-      )}
-      {showUnpublishModal && (
-        <ConfirmUnpublish
-          fileName={fileName}
-          setShowModal={setShowUnpublishModal}
-          onConfirm={() => handleUnpublish(fileName)}
-        />
-      )}
       {showPublishModal && (
         <Publish fileName={fileName} setShowModal={setShowPublishModal} />
       )}
