+++ src/components/SideBar.tsx
@@ -1,5 +1,9 @@
+import { useState, useCallback } from 'react'
 import { Link, useMatch } from 'react-router-dom'
+import { TrashIcon, ArchiveBoxXMarkIcon } from '@heroicons/react/24/outline'
 import { useStore } from '../state/base'
+import { ConfirmDeleteDraft, ConfirmUnpublish } from './Modal'
+import { api } from '../state/api'
 
 interface SideBarProps {
   onToggle: any
@@ -9,7 +13,46 @@ export default function SideBar({ onToggle }: SideBarProps) {
   const drafts = useStore((state) => state.drafts)
   const pages = useStore((state) => state.pages)
   const themes = useStore((state) => state.themes)
+  const { getAll } = useStore()
+  const [fileName, setFileName] = useState('')
+  const [showDeleteDraftModal, setShowDeleteDraftModal] = useState(false)
+  const [showUnpublishModal, setShowUnpublishModal] = useState(false)
   const match = useMatch('*')
+
+  const handleDeleteDraft = useCallback(async (toDelete: string) => {
+    await api.poke({
+      app: 'blog',
+      mark: 'blog-action',
+      json: {
+        'delete-draft': {
+          path: toDelete,
+        },
+      },
+    })
+    setShowDeleteDraftModal(false)
+    getAll()
+  }, [])
+
+  const handleUnpublish = useCallback(async (toUnpublish: string) => {
+    await api.poke({
+      app: 'blog',
+      mark: 'blog-action',
+      json: { unpublish: { path: toUnpublish } },
+    })
+    setShowUnpublishModal(false)
+    getAll()
+  }, [])
+
+  const showModal = (linkbase: string) => {
+    switch (linkbase) {
+      case '/published':
+        return setShowUnpublishModal(true)
+      case '/draft':
+        return setShowDeleteDraftModal(true)
+      default:
+        return
+    }
+  }
 
   type SidebarItemProps = {
     linkbase: string
@@ -19,17 +62,32 @@ export default function SideBar({ onToggle }: SideBarProps) {
   const SidebarItem = ({ linkbase, item }: SidebarItemProps) => {
     let linkto = `${linkbase}${item}`
     return (
-      <Link to={linkto}>
-        <li
-          className={`flex mb-1 text-blue-600 visited:text-purple-600 text-xs hover:bg-gray-100 p-2 ${
-            match?.pathname === linkto ? 'bg-gray-100' : ''
-          }`}
-        >
+      <li
+        className={`flex flex-row justify-between items-center mb-1 text-blue-600 visited:text-purple-600 text-xs hover:bg-gray-100 p-2 ${
+          match?.pathname === linkto ? 'bg-gray-100' : ''
+        }`}
+      >
+        <Link to={linkto}>
           <div className='text-left flex-1 my-auto truncate'>
             <code>{item}</code>
           </div>
-        </li>
-      </Link>
+        </Link>
+        {linkbase !== '/theme/' && (
+          <div
+            className='w-6 p-1 cursor-pointer rounded-sm text-red-500 hover:text-white hover:bg-red-500 '
+            onClick={() => {
+              showModal(linkbase)
+              setFileName(item)
+            }}
+          >
+            {linkbase === '/published' ? (
+              <ArchiveBoxXMarkIcon />
+            ) : (
+              <TrashIcon />
+            )}
+          </div>
+        )}
+      </li>
     )
   }
 
@@ -72,6 +130,20 @@ export default function SideBar({ onToggle }: SideBarProps) {
           <code>%new-theme</code>
         </button>
       </Link>
+      {showDeleteDraftModal && (
+        <ConfirmDeleteDraft
+          fileName={fileName}
+          setShowModal={setShowDeleteDraftModal}
+          onConfirm={() => handleDeleteDraft(fileName)}
+        />
+      )}
+      {showUnpublishModal && (
+        <ConfirmUnpublish
+          fileName={fileName}
+          setShowModal={setShowUnpublishModal}
+          onConfirm={() => handleUnpublish(fileName)}
+        />
+      )}
     </div>
   )
 }
